+++ core/src/com/artlessavian/subpar/states/CrouchState.java
@@ -60,7 +60,7 @@ public class CrouchState extends GroundedActionableState
 	}
 
 	@Override
-	public void act(float delta)
+	public void act(int delta)
 	{
 		super.act(delta);
