+++ core/src/com/artlessavian/subpar/states/EmptyState.java
@@ -23,7 +23,7 @@ public class EmptyState extends State
 	}
 
 	@Override
-	public void act(float delta)
+	public void act(int delta)
 	{
 	}
