+++ core/src/com/artlessavian/subpar/states/NormalState.java
@@ -6,7 +6,7 @@ import com.badlogic.ashley.core.Entity;
 
 public class NormalState extends AttackState
 {
-	float timer;
+	int timer;
 	final Entity entity;
 
 	public NormalState(Entity e)
@@ -57,7 +57,7 @@ public class NormalState extends AttackState
 	}
 
 	@Override
-	public void act(float delta)
+	public void act(int delta)
 	{
 		super.act(delta);
