+++ core/src/com/artlessavian/subpar/states/RunState.java
@@ -66,7 +66,7 @@ public class RunState extends GroundedActionableState
 	}
 
 	@Override
-	public void act(float delta)
+	public void act(int delta)
 	{
 		super.act(delta);
